chore: Use `gas` instead of `gasLimit` in e2e tests
This uses the `gas` parameter instead of `gasLimit` as we're
constructing an EIP1559 transaction (not a legacy tx).

This will ensure that for the policy evaluation SendEvmTransaction
tests that even though we have Insufficient Balance to send 0.1 ETH,
since we're specifying the `gas` param, we do not need to estimate
gas (a call that would fail for insufficient balance).